requirements.txt
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astor==0.8.1
async-generator==1.10
attrs==20.3.0
awsebcli==3.19.2
backcall==0.2.0
base58==2.0.1
bcrypt==3.2.0
bleach==3.2.1
blessed==1.17.12
blinker==1.4
botocore==1.19.35
cached-property==1.5.2
cachetools==4.2.0
cement==2.8.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.4
docker==4.4.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
future==0.16.0
gitdb==4.0.5
GitPython==3.1.11
icu==0.0.1
idna==2.10
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
Morfessor==2.0.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.1
pathspec==0.5.9
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
polyglot==16.7.4
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyarrow==2.0.0
pycld2==0.41
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
regex==2020.11.13
requests==2.24.0
scipy==1.5.4
semantic-version==2.5.0
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
spicy==0.16.0
sqlparse==0.4.1
stevedore==3.2.2
streamlit==0.72.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
texttable==1.6.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.25.11
validators==0.18.1
virtualenv==20.0.23
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
watchdog==1.0.1
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
wordcloud==1.8.1
